reservation-webapp/booking/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.3
branca==0.6.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.2
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.2
cssselect==1.2.0
customtkinter==5.1.2
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.7
Django==4.2.2
django-admin-black==1.0.7
dnspython==2.3.0
docx==0.2.4
et-xmlfile==1.1.0
exception==0.1.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.4
Fiona==1.9.1
Flask==2.3.3
Flask-MySQLdb==1.0.1
folium==0.14.0
fonttools==4.38.0
frozendict==2.3.7
future==0.18.3
geocoder==1.38.1
geojson==3.0.1
geopandas==0.12.2
googlemaps==4.10.0
grpcio==1.58.0
grpcio-tools==1.58.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.21.3
ipython==8.11.0
ipywidgets==8.0.4
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kivy-garden.mapview==1.0.6
kiwisolver==1.4.4
lxml==4.9.2
mapview==1.0.6
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
multitasking==0.0.11
munch==2.5.0
mutagen==1.46.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.0
osmnx==1.3.0
outcome==1.2.0
packaging==23.1
pandas==1.5.3
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.9.8
platformdirs==3.11.0
polyline==2.0.0
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
PyDispatcher==2.0.7
pygame==2.4.0
Pygments==2.14.0
pymongo==4.3.3
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
pyproj==3.4.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt6==6.5.0
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2022.7.1
pyzmq==25.0.0
queuelib==1.6.2
ratelim==0.1.6
requests==2.28.2
requests-file==1.5.1
scikit-learn==1.2.1
scipy==1.10.1
Scrapy==2.9.0
selenium==4.9.1
service-identity==23.1.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
threadpoolctl==3.1.0
tkintermapview==1.28
tldextract==3.4.4
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==1.26.14
virtualenv==20.24.5
virtualenv-clone==0.5.7
w3lib==2.1.1
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
whitenoise==6.5.0
widgetsnbextension==4.0.5
wsproto==1.2.0
xlrd==2.0.1
yfinance==0.2.18
yt-dlp==2023.3.4
zope.interface==6.0
